Refactor code_parser.py and test_custom_component.py
